Initial commit with full sample code
This commit includes the code for the fully working example. It has
three components: the app itself, the sun image wrapper and the color
selector.

The project has been divided in different folders, so tests, stylesheets
and main javascript code are not all mixed up in the same place.

The README.md is also included with generic instructions on how to run a
React application.